refactor(auth): expose facade over internal packages
