+++ packages/auth/internal/token/admin.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/golang-jwt/jwt/v5"
 
-	"github.com/e2b-dev/infra/packages/auth/pkg/token/jwks"
+	"github.com/e2b-dev/infra/packages/auth/internal/token/jwks"
 )
 
 // adminJWTClockSkew is the leeway applied to time-based claims of admin
