+++ packages/dashboard-api/internal/cfg/model.go
@@ -6,17 +6,17 @@ import (
 
 	"github.com/caarlos0/env/v11"
 
-	"github.com/e2b-dev/infra/packages/auth/pkg/token"
+	sharedauth "github.com/e2b-dev/infra/packages/auth/pkg/auth"
 )
 
 type Config struct {
-	Port                        int                  `env:"PORT"                                         envDefault:"3010"`
-	PostgresConnectionString    string               `env:"POSTGRES_CONNECTION_STRING,required,notEmpty"`
-	ClickhouseConnectionString  string               `env:"CLICKHOUSE_CONNECTION_STRING"`
-	ClickhouseConnectionStrings []string             `env:"CLICKHOUSE_CONNECTION_STRINGS"                envSeparator:";"`
-	AdminToken                  string               `env:"ADMIN_TOKEN,required,notEmpty"`
-	AuthProvider                token.ProviderConfig `env:"AUTH_PROVIDER_CONFIG"`
-	AdminAuth                   token.ProviderConfig `env:"ADMIN_AUTH_CONFIG"`
+	Port                        int                       `env:"PORT"                                         envDefault:"3010"`
+	PostgresConnectionString    string                    `env:"POSTGRES_CONNECTION_STRING,required,notEmpty"`
+	ClickhouseConnectionString  string                    `env:"CLICKHOUSE_CONNECTION_STRING"`
+	ClickhouseConnectionStrings []string                  `env:"CLICKHOUSE_CONNECTION_STRINGS"                envSeparator:";"`
+	AdminToken                  string                    `env:"ADMIN_TOKEN,required,notEmpty"`
+	AuthProvider                sharedauth.ProviderConfig `env:"AUTH_PROVIDER_CONFIG"`
+	AdminAuthProvider           sharedauth.ProviderConfig `env:"ADMIN_AUTH_PROVIDER_CONFIG"`
 
 	AuthDBConnectionString            string `env:"AUTH_DB_CONNECTION_STRING"`
 	AuthDBReadReplicaConnectionString string `env:"AUTH_DB_READ_REPLICA_CONNECTION_STRING"`
@@ -75,8 +75,8 @@ func Parse() (Config, error) {
 	var config Config
 	err := env.ParseWithOptions(&config, env.Options{
 		FuncMap: map[reflect.Type]env.ParserFunc{
-			reflect.TypeFor[token.ProviderConfig](): func(v string) (any, error) {
-				return token.ParseProviderConfig(v)
+			reflect.TypeFor[sharedauth.ProviderConfig](): func(v string) (any, error) {
+				return sharedauth.ParseProviderConfig(v)
 			},
 		},
 	})
