+++ packages/dashboard-api/internal/identity/issuer_test.go
@@ -5,14 +5,14 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/e2b-dev/infra/packages/auth/pkg/token/jwks"
+	sharedauth "github.com/e2b-dev/infra/packages/auth/pkg/auth"
 )
 
 func TestResolveOryIssuer_SingleJWT(t *testing.T) {
 	t.Parallel()
 
-	issuer, err := ResolveOryIssuer("https://tenant.projects.oryapis.com", []jwks.Config{
-		{Issuer: jwks.Issuer{URL: "https://auth.example.com"}},
+	issuer, err := ResolveOryIssuer("https://tenant.projects.oryapis.com", []sharedauth.JWTConfig{
+		{Issuer: sharedauth.JWTIssuer{URL: "https://auth.example.com"}},
 	})
 	require.NoError(t, err)
 	require.Equal(t, "https://auth.example.com", issuer)
@@ -21,9 +21,9 @@ func TestResolveOryIssuer_SingleJWT(t *testing.T) {
 func TestResolveOryIssuer_MultipleJWTMatchesSDKHost(t *testing.T) {
 	t.Parallel()
 
-	issuer, err := ResolveOryIssuer("https://tenant.projects.oryapis.com", []jwks.Config{
-		{Issuer: jwks.Issuer{URL: "https://auth-a.mycompany.com"}},
-		{Issuer: jwks.Issuer{URL: "https://tenant.projects.oryapis.com"}},
+	issuer, err := ResolveOryIssuer("https://tenant.projects.oryapis.com", []sharedauth.JWTConfig{
+		{Issuer: sharedauth.JWTIssuer{URL: "https://auth-a.mycompany.com"}},
+		{Issuer: sharedauth.JWTIssuer{URL: "https://tenant.projects.oryapis.com"}},
 	})
 	require.NoError(t, err)
 	require.Equal(t, "https://tenant.projects.oryapis.com", issuer)
@@ -32,9 +32,9 @@ func TestResolveOryIssuer_MultipleJWTMatchesSDKHost(t *testing.T) {
 func TestResolveOryIssuer_MultipleJWTNoMatch(t *testing.T) {
 	t.Parallel()
 
-	_, err := ResolveOryIssuer("https://tenant.projects.oryapis.com", []jwks.Config{
-		{Issuer: jwks.Issuer{URL: "https://auth-a.mycompany.com"}},
-		{Issuer: jwks.Issuer{URL: "https://auth-b.mycompany.com"}},
+	_, err := ResolveOryIssuer("https://tenant.projects.oryapis.com", []sharedauth.JWTConfig{
+		{Issuer: sharedauth.JWTIssuer{URL: "https://auth-a.mycompany.com"}},
+		{Issuer: sharedauth.JWTIssuer{URL: "https://auth-b.mycompany.com"}},
 	})
 	require.Error(t, err)
 	require.Contains(t, err.Error(), "no JWT issuer")
@@ -59,9 +59,9 @@ func TestResolveOryIssuer_NoJWTConfigsAndNoSDKURL(t *testing.T) {
 func TestResolveOryIssuer_DeduplicatesSameIssuer(t *testing.T) {
 	t.Parallel()
 
-	issuer, err := ResolveOryIssuer("https://tenant.projects.oryapis.com", []jwks.Config{
-		{Issuer: jwks.Issuer{URL: "https://auth.example.com"}},
-		{Issuer: jwks.Issuer{URL: "https://auth.example.com"}},
+	issuer, err := ResolveOryIssuer("https://tenant.projects.oryapis.com", []sharedauth.JWTConfig{
+		{Issuer: sharedauth.JWTIssuer{URL: "https://auth.example.com"}},
+		{Issuer: sharedauth.JWTIssuer{URL: "https://auth.example.com"}},
 	})
 	require.NoError(t, err)
 	require.Equal(t, "https://auth.example.com", issuer)
